entra-agent-cli/requirements.txt
typer[all]>=0.9.0
msal>=1.24.0
httpx>=0.25.0
python-dotenv>=1.0.0
rich>=13.0.0

